evolution plugin moved into client/evolution for a correct and extensible naming convention.
EcDoRpc IDL modified:
- opnum are now 8 bits
- action enum added 
- EcDoRpc ndrdump output enhanced

IMAPISession IDL modified:
- Change OpenMsgStore function name to MAPI_RPC_LOGON
for the 0xFE opnum operation.

Fix warnings in the code.

jkerihuel.

+++ libmapi/emsmdb.c
@@ -103,9 +103,9 @@ struct emsmdb_context *emsmdb_connect(TALLOC_CTX *mem_ctx, struct dcerpc_pipe *p
 	r.in.alloc_space = talloc(mem_ctx, uint32_t);
 	*r.in.alloc_space = 0;
 
-	r.out.unknown4[0] = talloc(mem_ctx, uint32_t);
-	r.out.unknown4[1] = talloc(mem_ctx, uint32_t);
-	r.out.unknown4[2] = talloc(mem_ctx, uint32_t);
+	r.out.unknown4[0] = (uint32_t)talloc(mem_ctx, uint32_t);
+	r.out.unknown4[1] = (uint32_t)talloc(mem_ctx, uint32_t);
+	r.out.unknown4[2] = (uint32_t)talloc(mem_ctx, uint32_t);
 	r.out.session_nb = talloc(mem_ctx, uint16_t);
 	r.out.alloc_space = talloc(mem_ctx, uint32_t);
 	r.out.handle = &ret->handle;
